backend/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h11==0.11.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
python-multipart==0.0.5
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
starlette==0.13.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1